fix(type): move ThumbnailSizeInPackedItem in thumbnailSizes to avoid any
