+++ src/enums/thumbnailSizes.ts
@@ -10,3 +10,13 @@ export const ThumbnailSize = {
   Original: 'original',
 } as const;
 export type ThumbnailSizeType = UnionOfConst<typeof ThumbnailSize>;
+
+// This type is defined here in the same file as ThumbnailSize.
+// Indeed, if defined in PackedItem, the transpiled type contains any.
+export const ThumbnailSizeInPackedItem = {
+  [ThumbnailSize.Small]: ThumbnailSize.Small,
+  [ThumbnailSize.Medium]: ThumbnailSize.Medium,
+} as const;
+export type ThumbnailsBySize = {
+  [thumbnailSize in UnionOfConst<typeof ThumbnailSizeInPackedItem>]: string;
+};
